Hp 장애물 충돌 시 튕김 수정

+++ Assets/Scripts/Items/CPItem.cs
@@ -21,7 +21,7 @@ public class CPItem : Item
 
     IEnumerator IncreaseCP(PlayerController player)
     {
-        int playerAtk = player.atk;
+        float playerAtk = player.atk;
         player.atk += CP;
         
         yield return new WaitForSeconds(plusCPTime);
